hw4/requirements.txt
pyramid==1.10
pyramid-jinja2==2.7
requests==2.18
mysql-connector-python
python-dotenv